updatecommit

+++ src/basic/dev/Shape.java
@@ -2,6 +2,6 @@ package basic.dev;
 
 public interface Shape {
 	String name();
-	float chuVi();
-	float dienTich();
+	double chuVi();
+	double dienTich();
 }
